agent-engine/requirements.txt
fastapi
uvicorn
google-generativeai
weave
python-dotenv
requests
